requirements.txt
discord.py~=2.2.3
Pillow~=9.5.0
DateTime~=5.1
python-dotenv~=1.0.0
pilmoji~=2.0.2
pathlib~=1.0.1